refactor: SearchBar.tsx 파일 이름이 변경되어 SearchBar 컴포넌트 import 수정
