+++ client/src/components/SearchBar/SearchBar.tsx
